pyproject.toml
[project]
name = "jerry-data-plane"
version = "0.1.0"
description = "Data plane for Jerry - Spark-based data processing and API server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "psycopg>=3.1.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "langchain==0.1.20",
    "langchain-openai==0.1.6",
    "langchain-community==0.0.38",
    "httpx>=0.25.0",
    "transformers>=4.57.3",
    "torch>=2.0.0",
]


[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]


